chore: remove original `SlurCard`
